+++ lib/namespace/mod.ts
@@ -1,5 +1,6 @@
 export * from "./model.ts";
 export * from "./interfaces.ts";
 export * from "./memory-repository.ts";
+export * from "./kv-repository.ts";
 export * from "./service.ts";
 export * from "./conformance.ts";